Cleaned up tests.
