+++ PapyrusCore/Boundary.swift
@@ -1,6 +1,6 @@
 //
 //  Boundary.swift
-//  Papyrus
+//  PapyrusCore
 //
 //  Created by Chris Nevin on 14/08/2015.
 //  Copyright © 2015 CJNevin. All rights reserved.
