+++ PapyrusCore/Papyrus.swift
@@ -1,6 +1,6 @@
 //
 //  Papyrus.swift
-//  Papyrus
+//  PapyrusCore
 //
 //  Created by Chris Nevin on 8/07/2015.
 //  Copyright © 2015 CJNevin. All rights reserved.
