+++ PapyrusCore/Position.swift
@@ -1,6 +1,6 @@
 //
 //  Position.swift
-//  Papyrus
+//  PapyrusCore
 //
 //  Created by Chris Nevin on 14/08/2015.
 //  Copyright © 2015 CJNevin. All rights reserved.
