+++ PapyrusCoreTests/PlayerTests.swift
@@ -1,6 +1,6 @@
 //
 //  PlayerTests.swift
-//  Papyrus
+//  PapyrusCore
 //
 //  Created by Chris Nevin on 13/09/2015.
 //  Copyright © 2015 CJNevin. All rights reserved.
